feat(sb-utils): create common stories for AI-components group
