+++ packages/plasma-new-hope/src/examples/components/_ai/UserMessage/UserMessage.stories.tsx
@@ -1,237 +1,25 @@
-import React, { ComponentProps, useRef, useState } from 'react';
-import { disableProps, getConfigVariations } from '@salutejs/plasma-sb-utils';
-import type { StoryObj, Meta } from '@storybook/react-vite';
-import { action } from 'storybook/actions';
+import type { ComponentProps } from 'react';
+import type { Meta } from '@storybook/react-vite';
+import { getUserMessageStories } from '@salutejs/plasma-sb-utils';
 
 import { WithTheme } from '../../../_helpers';
-import { IconEditOutline, IconPlasma } from '../../../../components/_Icon';
 import { Button } from '../../Button/Button';
 import { EmbedIconButton } from '../../EmbedIconButton/EmbedIconButton';
 
 import { UserMessage } from './UserMessage';
 import { config } from './UserMessage.config';
 
-const { views, sizes } = getConfigVariations(config);
-const fileContentViews = ['default', 'secondary'];
-const attachmentTypes = ['file', 'fileGroup', 'image', 'none'] as const;
-
-const onEdit = action('onEdit');
-const onEditKeyDown = action('onEditKeyDown');
-const onFocus = action('onFocus');
-const onBlur = action('onBlur');
-
-type StoryProps = ComponentProps<typeof UserMessage> & {
-    attachmentType?: typeof attachmentTypes[number];
-    initialValue?: string;
-};
-
-type AttachmentsMap = Record<typeof attachmentTypes[number], StoryProps['attachments']>;
-
-const attachmentsMap: AttachmentsMap = {
-    file: [
-        {
-            id: 'file-1',
-            type: 'file',
-            label: 'document.pdf',
-            size: '2.4 МБ',
-        },
-    ],
-    fileGroup: [
-        {
-            id: 'file-1',
-            type: 'file',
-            label: 'report.pdf',
-            size: '2.4 МБ',
-        },
-        {
-            id: 'file-2',
-            type: 'file',
-            label: 'very long named document with time stamp.doc',
-            size: '2.4 МБ',
-        },
-        {
-            id: 'file-3',
-            type: 'file',
-            label: 'data.csv',
-            size: '2.4 МБ',
-        },
-    ],
-    image: [
-        {
-            id: 'file-5',
-            type: 'image',
-            label: 'картинка для примера фоном',
-            size: '123 КБ',
-            thumbUrl: './images/320_320_9.jpg',
-        },
-    ],
-    none: undefined,
-};
-
-const meta: Meta<StoryProps> = {
-    title: 'AI Kit/UserMessage',
+type UserMessageProps = ComponentProps<typeof UserMessage>;
+const { meta: META, Default } = getUserMessageStories({
     component: UserMessage,
+    componentConfig: config,
+    Button,
+    EmbedIconButton,
     decorators: [WithTheme],
-    parameters: {
-        controls: {
-            disableSaveFromUI: true,
-        },
-    },
-    argTypes: {
-        view: {
-            description: 'Вид компонента',
-            options: views,
-            control: { type: 'select' },
-            table: { category: 'variation' },
-        },
-        size: {
-            description: 'Размер компонента',
-            options: sizes,
-            control: { type: 'select' },
-            table: { category: 'variation' },
-        },
-        attachmentsView: {
-            description: 'Вид прикреплённых файлов',
-            options: fileContentViews,
-            control: { type: 'select' },
-            table: { category: 'variation' },
-        },
-        maxRows: {
-            description: 'Максимальное количество строк при редактировании',
-            control: { type: 'number' },
-            table: { category: 'layout-related' },
-        },
-        attachmentType: {
-            description: 'Тип вложений для демонстрации',
-            options: attachmentTypes,
-            control: { type: 'select' },
-            table: { category: 'content-related' },
-        },
-        initialValue: {
-            description: 'Начальное значение текста сообщения',
-            control: 'text',
-            table: { category: 'content-related' },
-        },
-        ...disableProps([
-            'actions',
-            'attachments',
-            'isEditing',
-            'onEdit',
-            'onEditKeyDown',
-            'value',
-            'textAreaRef',
-            'onFocus',
-            'onBlur',
-            'onEditKeyDown',
-        ]),
-    },
-    args: {
-        initialValue: 'Привет! Что насчет того, чтобы пройти тест Тьюринга?',
-        view: 'default',
-        size: 'm',
-        attachmentsView: 'default',
-        attachmentType: 'fileGroup',
-        maxRows: 4,
-    },
+});
+const meta: Meta<UserMessageProps> = {
+    ...META,
+    title: 'AI Kit/UserMessage',
 };
-
 export default meta;
-
-const StoryDefault = (args: StoryProps) => {
-    const [value, setValue] = useState(args.initialValue);
-    const [editingValue, setEditingValue] = useState(args.initialValue);
-    const [isEditing, setIsEditing] = useState(false);
-
-    const textAreaRef = useRef<HTMLTextAreaElement>(null);
-
-    const handleEditBegin = () => {
-        setEditingValue(value);
-        setIsEditing(true);
-
-        setTimeout(() => {
-            if (textAreaRef.current) {
-                textAreaRef.current.focus();
-                textAreaRef.current.setSelectionRange(value?.length || 0, value?.length || 0);
-            }
-        });
-    };
-
-    const handleEditComplete = () => {
-        setValue(editingValue);
-        setIsEditing(false);
-    };
-
-    const renderActions = () => {
-        const buttonSize = () => {
-            switch (args.size) {
-                case 'xl':
-                case 'l':
-                    return 'm';
-
-                case 'm':
-                    return 's';
-
-                case 's':
-                    return 'xs';
-
-                case 'xs':
-                    return 'xxs';
-
-                default:
-                    return 's';
-            }
-        };
-        const iconSize = args.size === 'xs' || args.size === 's' ? 'xs' : 's';
-
-        if (isEditing) {
-            return (
-                <>
-                    <Button view="default" size={buttonSize()} onClick={() => setIsEditing(false)}>
-                        Отмена
-                    </Button>
-                    <Button view="accent" size={buttonSize()} onClick={handleEditComplete}>
-                        Сохранить
-                    </Button>
-                </>
-            );
-        }
-
-        return (
-            <>
-                {Array.from({ length: 3 }).map((_, index) => (
-                    <EmbedIconButton size="m" view="secondary" key={index}>
-                        <IconPlasma size={iconSize} color="inherit" />
-                    </EmbedIconButton>
-                ))}
-
-                <EmbedIconButton size="m" view="secondary" onClick={handleEditBegin}>
-                    <IconEditOutline size={iconSize} color="inherit" />
-                </EmbedIconButton>
-            </>
-        );
-    };
-
-    return (
-        <div style={{ display: 'flex', justifyContent: 'flex-end', padding: '1rem' }}>
-            <UserMessage
-                {...args}
-                textAreaRef={textAreaRef}
-                attachments={attachmentsMap[args.attachmentType as keyof typeof attachmentsMap]}
-                value={isEditing ? editingValue : value}
-                actions={renderActions()}
-                isEditing={isEditing}
-                onEdit={(e) => {
-                    setEditingValue(e.target.value);
-                    onEdit(e);
-                }}
-                onEditKeyDown={onEditKeyDown}
-                onFocus={onFocus}
-                onBlur={onBlur}
-            />
-        </div>
-    );
-};
-
-export const Default: StoryObj<StoryProps> = {
-    render: StoryDefault,
-};
+export { Default };
