+++ utils/plasma-sb-utils/src/components/_ai/UserMessage/index.ts
@@ -1,0 +1,1 @@
+export * from './UserMessage';
