+++ utils/plasma-sb-utils/src/components/index.ts
@@ -42,3 +42,4 @@ export * from './Badge';
 export * from './CodeField';
 export * from './CodeInput';
 export * from './CodeArea';
+export * from './_ai';
